Refactor case folding methods and improve Unicode handling
- Renamed parameters in caseFoldSimple methods for clarity.
- Updated case folding logic to use 'intervals' instead of 'ranges' in ClassUnicodeRange and ClassBytesRange.
- Removed unused typealias for UnicodeRangeTable.
- Enhanced comments in generated property names file to include generation details.
- Added comprehensive tests for Unicode functionalities, including case folding and symbolic name normalization.

+++ src/commonMain/kotlin/io/github/kotlinmania/regexsyntax/parser/Parser.kt
@@ -7,8 +7,8 @@ package io.github.kotlinmania.regexsyntax.parser
  */
 
 import io.github.kotlinmania.regexsyntax.Error
-import io.github.kotlinmania.regexsyntax.ast.parse.AstParserBuilder
-import io.github.kotlinmania.regexsyntax.ast.parse.AstParser
+import io.github.kotlinmania.regexsyntax.ast.parse.ParserBuilder as AstParserBuilder
+import io.github.kotlinmania.regexsyntax.ast.parse.Parser as AstParser
 import io.github.kotlinmania.regexsyntax.hir.Hir
 import io.github.kotlinmania.regexsyntax.hir.translate.Translator
 import io.github.kotlinmania.regexsyntax.hir.translate.TranslatorBuilder
@@ -77,7 +77,7 @@ data class ParserBuilder(
      * used in a granular way.
      */
     fun nestLimit(limit: Int): ParserBuilder {
-        ast.nestLimit(limit)
+        ast.nestLimit(limit.toUInt())
         return this
     }
